moved packaging of dbg repo out of cli
